add python prototype of lexer
This works by transferring the first character from the input string to
the string buffer and detecting if it matches a token type based on
regex. Ff the string buffer matches an error, it outputs the last
correct match it had, puts the character that caused the error back into
the input string, and clears the string buffer. If it had no correct
match and immediately errors, it is an unrecognized symbol and outputs
an error token. The character is NOT put back into the input string and
the string buffer is cleared. There are still bugs in this prototype
with edge cases and ordering of the tokens in the input string.