make parser be able to construct a parse tree
This is done by having a node passed into each production. A production
also holds a "local node" in which children such as other nodes and
terminals get added into. Other nodes get added if the production travels
to a non-terminal. If the string matches the production, it becomes a
child of the node passed in.