"recursive descent" PEG parser
Normally, left-associative operators are written left-recursively in the
grammar. However, recursive descent cannot handle left recursion so the
CFG is rewritten to become a Parsing Expression Grammar (PEG). It
combines some regex symbols, allowing productions such as `A: M ('+' M)*`.
This production can be flattened by using while loops, not having to use
recursion. The recursive descent structure is still present with how it
handles the operator precedence and when a non-terminal has multiple
productions. The next parser, Pratt parser, will handle operator
precedence and associativity using tables instead of taking advantage of
the structure of the grammar.