solve infinite recursion problem by restructuring productions and nonterminal functions
Nonterminal functions now just consists of rules, a set of terminals and
nonterminals, which are passed into `parse_productions()`. This function
iterates through the rules calls `production()` for each rule.
`production()` parses the terminals and nonterminals like normal. If a
terminal or nonterminal fails, returns `None`, the function early
returns `None` and fails, fixing the infinite recursion problem. This
also makes it easier to handwrite nonterminal functions and its
productions.