add `if` statements and `blocks` to pratt parser. fix else in grammar.
the pratt parser does not handle if statements and blocks directly. it
instead redirects them back to the packrat parser. this is because
blocks can contain statements which include declarations, assignments,
and while tatements, which are not considered expressions.

else statements now enforces blocks. to be able to have else-if
statements, if statements are added to else statement's productions
directly.